requirements.txt
Django==1.8.3
argparse==1.2.1
django-modeltranslation==0.10
wsgiref==0.1.2
